changed int by float

+++ Assets/Scripts/Collectables/RotateObject.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 public class RotateObject : MonoBehaviour
 {
-    public int rotateSpeed = 1;
+    public float rotateSpeed = 1;
 
     void Update()
     {
